Automated rollback of changelist.
*** Reason for rollback ***

Second attempt at checking in the auto escaping code. All the htmlparser fixes and
changes went into the previous CL, there are no changes to htmlparser in this CL.
This is essentially a rollback of the previous rollback with
- the new htmlparser naming conventions HTML_STATE ==> HTMLPARSER_STATE, state ==> int
- make ContentTypeList in neo_auto.c static.
- Added libclearsilver-jni.so, neo_cgi.so. libclearsilver.jar did not change.

Adding auto escaping to clearsilver. util/neos_auto.c and util/neos_auto.h contain the actual escaping logic. htmlparser/* contains a preliminary version of the HTML parsing code.