Cloned from mugdha@mbendre on 2010/06/23 11:49:27
Fix a problem with url validation. Previously in order to extract the
URI scheme out of a URL we were looking for a ':' anywhere in the URL as
long as there wasn't a '/' before it. However, this fails for a URL with
no scheme but containing an unencoded ':' in a query parameter. for
instance Login?continue=http://www.google.com - the whole url up to the
':' was treated as the URI scheme.

Now we apply the syntax defined by RFC 3986 for a valid scheme and only
accept alphanumeric, +, -, . as valid characters in the scheme.