Clone from mugdha@mbendre-auto-escape on 2011/08/01 12:54:05
Fix url validation bypass by appending whitespace to beginning of
variable. eg " javascript:alert(1)".

This xss was introduced because the validation method was changed in
response to another bug. Before that, the validation used to treat the
scheme as everything from start of string up to the first colon before
occurence of '/' (or just the first colon if there's no '/'.) This
caused a problem when there was a relative url with unencoded :// in a
query parameter.
In response to that, I had changed the code to its current incarnation
of only treating alphanumeric characters in the scheme.

In this CL I go back to the previous design, which seems like it was the
safer design - it made less assumptions about what constitutes a
URI scheme. Just, instead of ignoring any : that occurs after
the occurence of '/', I ignore any : that occurs after occurence
of '/' or ':'.