Cloned from blong@clearsilver on 2017/06/29 15:50:44
Always allocate the string in eval_expr_string when we have only a
single argument, since we can't guarantee the lifetime of the result
will be shorter than the source, since the source can be an hdf var
which can be modified and the result can be used as a local var.

Also fix a memory leak in the parser on invalid arguments to call found
in the same fuzzer run.